Deduplicate physics constants and simplify the training entry points
Share a single invariant_mass2 helper in physics.physics across the numpy
loader/selection call sites and the torch loss module, which each had their
own copy. Replace the remaining duplicated literals with the constants that
already existed: TOR for the 1e-16 guards in the loss module, the boosting
helper and the ONNX check, and HIGGS_MASS for the 125.0 in WConstraintsLayer
and the LightningWBoson default.

Flatten load_data's row filtering back into the loader instead of routing it
through _row_masks and _valid_dilepton_mass_rows, keeping the m_ll rationale
as a comment at the cut. The selection is unchanged: the new and old masks
agree row for row over all three ggF splits.

Drop the WBosonDataModule.setup override and its call sites. The splits are
built in __init__, so the override only shadowed the Lightning no-op.

In train.py, extract evaluate_best_checkpoint from run_training and read the
optional argparse namespace one consistent way. In k_fold_train.py, validate
--fold with an argparse type so it arrives as an int, and move the row-count
check to the pooled arrays it actually guards. run_k_fold_train.sh now passes
the default config directly, since a user-supplied --config later on the
command line already overrides it.

Rewrite the higgs_scale docstring around the stabilized form it evaluates,
and apply ruff format across the tree.

+++ physics/diagnostics.py
@@ -15,7 +15,7 @@ def component_calibration(prediction, truth):
         )
 
     metrics = {
-        name: np.full(prediction.shape[1], np.nan, dtype=np.float64)
+        name: np.full(prediction.shape[1], np.nan)
         for name in ("bias", "width_ratio", "correlation", "slope", "intercept", "rmse")
     }
     for index in range(prediction.shape[1]):
